fix: fixed for the issue (#248)
